use build service to synchronize access to ports between system tests
