+++ buildSrc/src/main/kotlin/com.teamscale.system-test-convention.gradle.kts
@@ -2,8 +2,11 @@ plugins {
 	id("com.teamscale.java-convention")
 }
 
+SystemTestPorts.registerWith(project)
+
 tasks.test {
 	dependsOn(":agent:shadowJar")
+	usesService(SystemTestPorts.provider)
 	systemProperties("agentPort" to agentPort, "teamscalePort" to teamscalePort)
 }
